Move check_{bids|caps}_folder (resp. fix_join) to inputs.py (resp. fix_nipype.py)
